launch 0.0.4

+++ up.sh
@@ -1,23 +1,38 @@
-#!/usr/bin/zsh -eux
+#!/usr/bin/zsh
 echo -n "plz version: "
 read -r v
+
 echo "new version: ${v}"
 echo -n "update files ok?:"
-read
-for i in README.md setup.py wbsv/__init__.py wbsv/Archive.py;do
-  sed -r "s/[0-9]\.[0-9]\.[0-9]/${v}/g" "${i}" |sponge "${i}"
-done
+read -r f
+if [[ "${f}" -eq "y" ]];then
+  for i in README.md setup.py wbsv/__init__.py wbsv/Archive.py;do
+    sed -r "s/[0-9]\.[0-9]\.[0-9]/${v}/g" "${i}" |sponge "${i}"
+  done
+fi
+
 echo -n "rm build dist wbsv.egg-info ok?:"
-read
-sudo rm -rf build dist wbsv.egg-info
-python3 setup.py sdist bdist_wheel
+read -r f
+if [[ "${f}" -eq "y" ]];then
+  sudo rm -rf build dist wbsv.egg-info
+  python3 setup.py sdist bdist_wheel
+fi
+
 echo -n "commit ok?:"
-read
-git add .
-git commit -m "launch ${v}"
+read -r f
+if [[ "${f}" -eq "y" ]];then
+  git add .
+  git commit -m "launch ${v}"
+fi
+
 echo -n "deploy pypi ok?:"
-read
-/home/eggplants/usr/pip/bin/twine upload --repository pypi dist/*
+read -r f
+if [[ "${f}" -eq "y" ]];then
+  twine upload --repository pypi dist/*
+fi
+
 echo -n "push github ok?:"
-read
-git push
+read -r f
+if [[ "${f}" -eq "y" ]];then
+  git push
+fi
